%  File networkDynamic/man/networkDynamic-internal.Rd
%  Part of the statnet package, http://statnetproject.org
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) in
%    http://statnetproject.org/attribution
%
%  Copyright 2012 the statnet development team
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{networkDynamic-internal}
\title{Internal networkDynamic objects}
\alias{.First.lib}
\alias{insert.spell}
\alias{delete.spell}
\description{Internal networkDynamic objects.}
\details{These are not to be called by the user.}
\keyword{internal}
